Polishing
- make ant dependency explicit
- share permission handling across unzip and untar transforms
