+++ buildSrc/src/main/java/org/elasticsearch/gradle/transform/UnpackTransform.java
@@ -31,8 +31,6 @@ import org.gradle.internal.UncheckedException;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 public interface UnpackTransform extends TransformAction<TransformParameters.None> {
 
@@ -43,43 +41,14 @@ public interface UnpackTransform extends TransformAction<TransformParameters.Non
     @Override
     default void transform(TransformOutputs outputs) {
         File archiveFile = getArchiveFile().get().getAsFile();
-        File unzipDir = outputs.dir(archiveFile.getName());
+        File extractedDir = outputs.dir(archiveFile.getName());
         try {
-            unpack(archiveFile, unzipDir);
+            unpack(archiveFile, extractedDir);
         } catch (IOException e) {
             throw UncheckedException.throwAsUncheckedException(e);
         }
     }
 
     void unpack(File archiveFile, File targetDir) throws IOException;
-
-    /*
-     * We want to remove up to the and including the jdk-.* relative paths. That is a JDK archive is structured as:
-     *   jdk-12.0.1/
-     *   jdk-12.0.1/Contents
-     *   ...
-     *
-     * and we want to remove the leading jdk-12.0.1. Note however that there could also be a leading ./ as in
-     *   ./
-     *   ./jdk-12.0.1/
-     *   ./jdk-12.0.1/Contents
-     *
-     * so we account for this and search the path components until we find the jdk-12.0.1, and strip the leading components.
-     */
-    static Path trimJdkArchiveExtractPath(String relativePath) {
-        final Path entryName = Paths.get(relativePath);
-        int index = 0;
-        for (; index < entryName.getNameCount(); index++) {
-            if (entryName.getName(index).toString().matches("jdk-?\\d.*")) {
-                break;
-            }
-        }
-        if (index + 1 >= entryName.getNameCount()) {
-            // this happens on the top-level directories in the archive, which we are removing
-            return null;
-        }
-        // finally remove the top-level directories from the output path
-        return entryName.subpath(index + 1, entryName.getNameCount());
-    }
 
 }
